+++ Nos/Service/Relay/RelaySubscriptionManager.swift
@@ -55,8 +55,6 @@ actor RelaySubscriptionManagerActor: RelaySubscriptionManager {
         all
     }
     
-    // MARK: - Init
-    
     private var socketQueue: DispatchQueue?
     private var delegate: WebSocketDelegate?
     
@@ -131,12 +129,7 @@ actor RelaySubscriptionManagerActor: RelaySubscriptionManager {
         all.forEach { relayAddresses.insert($0.relayAddress) }
 
         for relayAddress in relayAddresses {
-            let connection: WebSocketConnection
-            if let existingConnection = socketConnections[relayAddress] {
-                connection = existingConnection
-            } else {
-                connection = createSocket(for: relayAddress)
-            }
+            let connection = findOrCreateSocket(for: relayAddress)
             
             switch connection.state {
             case .errored(let error):
@@ -156,7 +149,11 @@ actor RelaySubscriptionManagerActor: RelaySubscriptionManager {
     
     /// Creates a WebSocketConnection for a relay. This is not idempotent - make sure it's called only once for 
     /// each relay.
-    private func createSocket(for relayAddress: URL) -> WebSocketConnection {
+    private func findOrCreateSocket(for relayAddress: URL) -> WebSocketConnection {
+        if let existingConnection = socketConnections[relayAddress] {
+            return existingConnection
+        } 
+        
         var request = URLRequest(url: relayAddress)
         request.timeoutInterval = 10
         let socket = WebSocket(request: request)
@@ -337,37 +334,4 @@ actor RelaySubscriptionManagerActor: RelaySubscriptionManager {
         }
         processSubscriptionQueue()
     }
-}
-
-/// A container that tracks how many times we have tried unsuccessfully to open a websocket and the next time we should
-/// try again.
-struct WebSocketErrorEvent: Equatable {
-    var retryCounter: Int = 1
-    var nextRetry: Date = .now
-    
-    mutating func trackRetry() {
-        self.retryCounter += 1
-        let delaySeconds = NSDecimalNumber(
-            decimal: pow(2, min(retryCounter, RelaySubscriptionManagerActor.maxBackoffPower))
-        )
-        self.nextRetry = Date(timeIntervalSince1970: Date.now.timeIntervalSince1970 + delaySeconds.doubleValue)
-    }
-}
-
-/// Represents a connection to a websocket with state tracking. Utilizes a Starscream WebSocket under the hood.
-class WebSocketConnection {
-    let socket: WebSocket
-    var state: WebSocketState 
-    
-    init(socket: WebSocket, state: WebSocketState = .disconnected) {
-        self.socket = socket
-        self.state = state
-    }
-}
-
-/// The states a WebSocketConnection can be in. These states are used by `RelaySubscriptionManager` to move each 
-/// socket to the `connected` state where we can start executing requests, taking into account time to connect, errors
-/// and authentication.
-enum WebSocketState: Equatable {
-    case disconnected, connecting, connected, errored(WebSocketErrorEvent)
 }
